Wound up having to make a lot of little changes (mainly to zip() and dict.keys() and .values() statements) in order to get the whole model working in Python 3, since I'll be presenting in that environment during lab meeting tomorrow. Also had to comment out a line added by Jason or Irene to mating.py in order to get the code to work.

+++ python_model/burn_in.py
@@ -23,7 +23,7 @@ def adf_threshold_test(pop, num_timesteps_back, p_val):
 
 def tt_threshold_test(pop, num_timesteps_back, p_val):
     num_timesteps_back += num_timesteps_back % 2
-    return(tt(pop.Nt[-num_timesteps_back: -num_timesteps_back/2], pop.Nt[-num_timesteps_back/2:])[1] >0.05)
+    return(tt(pop.Nt[int(-num_timesteps_back): int(-num_timesteps_back/2)], pop.Nt[int(-num_timesteps_back/2):])[1] >0.05)
